Add files via upload
CloudSim is the server side where multi clouds are periodically generated and so is their service files and services.

MultiCuckoo is the client side where the MultiCuckoo algorithm was developed and any other algorithms can be developed and tested. The server side should be running first. The user request is input from the console.